refactor(moonpool-core): split the storage module, document the file contract
`storage.rs` held three concerns in one file: the open options, the two
provider traits, and the production Tokio implementation. Split them into
`storage/{options,mod,tokio_impl}.rs` so each file holds one, and write down
the contract the rest of this series builds on:

- the provider owns the filesystem *namespace*, the file owns already-open
  bytes;
- a completed write is visible, never durable — only a sync is durable;
- `set_len` grows the addressable range, and is not preallocation, not
  initialization, and not a metadata barrier;
- there is one file image behind every access path, so a database engine can
  be written directly against `StorageFile` rather than against a second
  storage stack built beside it.

Adds the `OpenOptions::read_write()` preset (the shape an engine opens an
existing data file with) and unit tests for the flag round-trip; no
behavioural change.
