+++ crates/moonpool-core/src/storage/tokio_impl.rs
@@ -1,0 +1,123 @@
+//! Production [`StorageProvider`] / [`StorageFile`] over the real filesystem.
+
+use std::io::{self, SeekFrom};
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+use futures::io::{AsyncRead, AsyncSeek, AsyncWrite};
+use tokio_util::compat::{Compat, TokioAsyncReadCompatExt};
+
+use super::{OpenOptions, StorageFile, StorageProvider};
+
+/// Real Tokio storage implementation.
+#[derive(Debug, Clone, Default)]
+pub struct TokioStorageProvider;
+
+impl TokioStorageProvider {
+    /// Create a new Tokio storage provider.
+    #[must_use]
+    pub fn new() -> Self {
+        Self
+    }
+}
+
+impl StorageProvider for TokioStorageProvider {
+    type File = TokioStorageFile;
+
+    async fn open(&self, path: &str, options: OpenOptions) -> io::Result<Self::File> {
+        let file = tokio::fs::OpenOptions::new()
+            .read(options.is_read())
+            .write(options.is_write())
+            .create(options.is_create())
+            .create_new(options.is_create_new())
+            .truncate(options.is_truncate())
+            .append(options.is_append())
+            .open(path)
+            .await?;
+        Ok(TokioStorageFile {
+            inner: file.compat(),
+        })
+    }
+
+    async fn exists(&self, path: &str) -> io::Result<bool> {
+        tokio::fs::try_exists(path).await
+    }
+
+    async fn delete(&self, path: &str) -> io::Result<()> {
+        tokio::fs::remove_file(path).await
+    }
+
+    async fn rename(&self, from: &str, to: &str) -> io::Result<()> {
+        tokio::fs::rename(from, to).await
+    }
+}
+
+/// Wrapper for Tokio File to implement our trait.
+///
+/// Holds the underlying `tokio::fs::File` inside `tokio_util::compat::Compat`
+/// so the `futures::io` trait impls come through automatically. The four custom
+/// methods on [`StorageFile`] reach the inner [`tokio::fs::File`] via
+/// [`Compat::get_ref`] / [`Compat::get_mut`] to call tokio-specific APIs
+/// (`sync_all`, `sync_data`, `metadata`, `set_len`) that `Compat` itself
+/// does not expose.
+#[derive(Debug)]
+pub struct TokioStorageFile {
+    inner: Compat<tokio::fs::File>,
+}
+
+impl StorageFile for TokioStorageFile {
+    async fn sync_all(&self) -> io::Result<()> {
+        self.inner.get_ref().sync_all().await
+    }
+
+    async fn sync_data(&self) -> io::Result<()> {
+        self.inner.get_ref().sync_data().await
+    }
+
+    async fn size(&self) -> io::Result<u64> {
+        let metadata = self.inner.get_ref().metadata().await?;
+        Ok(metadata.len())
+    }
+
+    async fn set_len(&self, size: u64) -> io::Result<()> {
+        self.inner.get_ref().set_len(size).await
+    }
+}
+
+impl AsyncRead for TokioStorageFile {
+    fn poll_read(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &mut [u8],
+    ) -> Poll<io::Result<usize>> {
+        Pin::new(&mut self.inner).poll_read(cx, buf)
+    }
+}
+
+impl AsyncWrite for TokioStorageFile {
+    fn poll_write(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &[u8],
+    ) -> Poll<io::Result<usize>> {
+        Pin::new(&mut self.inner).poll_write(cx, buf)
+    }
+
+    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
+        Pin::new(&mut self.inner).poll_flush(cx)
+    }
+
+    fn poll_close(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
+        Pin::new(&mut self.inner).poll_close(cx)
+    }
+}
+
+impl AsyncSeek for TokioStorageFile {
+    fn poll_seek(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        pos: SeekFrom,
+    ) -> Poll<io::Result<u64>> {
+        Pin::new(&mut self.inner).poll_seek(cx, pos)
+    }
+}
